+++ CHANGELOG.md
@@ -10,6 +10,7 @@
 - **add:** joi based validation for route config objects
 - **add:** sinon (1.17.4) and sinon-chai (2.8.0)
 - **rmv:** do not lint test directory
+- **add:** functionality to log the route mapping
 
 ### [v0.2.1] - 2016-05-16
 - **doc:** fix typo